Make web_push_subscription.register an action
